fix styling, refactor
